Auto battrey reports for Windows Laptops
Developed in Windows 10, IT MAY NOT WORK ON WINDOWS 11 IF THERE ARE VARIOUS CHANGES. NEED TO TEST {^-^}

DATE FORMAT is "Thu 03.10.2024". Check your date format by typing "date" in CMD and adjust accordingly.
desktopPath is where the battery report and daily report will be saved.
You need to place this .bat file in the StartUp folder.
And of course, don’t forget to convert this to a .bat file. If this file send to you with .txt extension.
To do this, open File Explorer, click on 'View' in the top menu, and check the box for 'File name extensions' on the left.
This will allow you to see and change file extensions.
For the StartUp folder, press Win+R and type 'shell:startup'.

THIS APPLICATION WILL RUN EVERY TIME YOU BOOT UP THE COMPUTER. (☞ﾟヮﾟ)☞